Move autotag and clean options out of dialogs
